GKI: Support preservation of listed symbols
build_abi.sh: Add support for KMI_SYMBOL_LIST_ADD_ONLY and document
it. Setting this to 1 will cause --additions-only to be passed to
extract_symbols.

extract_symbols: If --additions-only is passed, the symbol list is
read before being overwritten and any symbols that would have been
lost in the new version are preserved at the end of the file.

Bug: 165839948
Change-Id: I3de4a916850d7543ec37ed92cb98c3f8039a6c07
Signed-off-by: Giuliano Procida <gprocida@google.com>