abi/download_from_ci: support downloading/creating KMI representations
When releasing GKI kernels it is useful to release the ABI/KMI
definition alongside the image. For that, add the functionality to
create a representation if needed or reuse from the build artefacts.
Distinguish this based on the abi.prop that is part of the build artefacts.
Use the existing abitool to actually do the creation if needed.

Bug: 166829073
Signed-off-by: Matthias Maennich <maennich@google.com>
Change-Id: I8ced7a4072a5970749590c4a0f566888c25966bb