build_abi: add support to use unstripped modules for ABI analysis
When the kernel modules under $DIST_DIR are stripped, allow the ABI
tooling to use the unstripped modules ($DIST_DIR/unstripped) to perform
the ABI analysis. This requires adding an additional parameter to
dump_api that specifies the vmlinux path since it would be different
from the kernel modules path (--linux-tree).

Bug: 156395749
Change-Id: I921d2b41fd9b1309f65611e57a476437bd3b3df6
Signed-off-by: Will McVicker <willmcvicker@google.com>