build: add support for mixed build mode
The mixed build mode allows to build a GKI kernel and a device kernel in
one run. The kernel builds share a DIST_DIR, hence vmlinux can be
provided by the GKI kernel and the modules can be provided by the device
kernel.

Note, the device kernel build's FILES variable must not contain the
kernel binary for that to work. For convenience, the variable
MIXED_BUILD is set for the device kernel config to detect that case.

Signed-off-by: Matthias Maennich <maennich@google.com>
Change-Id: Ie842d1ce2c17368c3e1fa8a68f3e4312ddb8930d