create_modules_staging: Correct grep usage in TRIM_UNUSED_MODULES
grep expects patterns to be prefixed with "-e" or to be in a list of
files with "-f" option. Presently, when modules.blocklist file is
included, the list of blocked modules is included directy in grep, e.g.
grep -f modules.order blocked_module.ko. This is not correct expected
arguments. A temporary file should be used to create proper list of
patterns for use by grep.

Change-Id: I21bce62313cc6255788fd34f655851fdbf0af2b7