hermetic builds: also strip away ${KERNEL_DIR}
Since ${KERNEL_DIR} is now dynamically defined, it might be
inconsistently defined in different contexts of the build. E.g.
common-android-multi vs. common-android12-5.10. This leads to
differently encoded filepaths in the debug info. A source of
indeterminism. It also confuses the ABI postprocessing as the full file
path is no longer present to replace.

Fix that by additionally strip the ${ROOT_DIR}/${KERNEL_DIR}. This
results in kernel sources being without any prefix and any other file
paths are prefixed with usually prebuilts/... (e.g. std header).

This works because -ffile-prefix-map can be passed multiple times and is
applied in the order of occurrence.

Bug: 180570392
Signed-off-by: Matthias Maennich <maennich@google.com>
Change-Id: I160d32dd3425108afe0818371627382ef24129db