build_abi.sh: Create a full ABI with no symbol list restriction
If no symbol list is applied at the point of ABI extraction, then the
full ABI is identical to that ABI. Otherwise, a full ABI is extracted
in parallel, with no symbol list.

The new files and symbolic links are named full-abi*.xml instead of
abi*.xml.

A full ABI, one unrestricted by the current symbol list, is useful for
various reasons.

- It allows vendors to check for ABI discrepancies using alternative
  symbol lists.
- It will allow us to verify that libabigail gives consistent results
  whether the symbol list is applied at extraction or diff time.
- It will allow us to review the kernel ABI over the history of a
  branch, independent of what may be happening with symbol lists.

Howver, if symbol lists are applied to exported symbols at build
time (via TRIM_NONLISTED_KMI), then this is all irrelevant.

Bug: 177882642
Change-Id: I934b1a18af1c54664082399adbd3a6635e4ffd74
Signed-off-by: Giuliano Procida <gprocida@google.com>