build: Add new LTO=[full|thin|none] option
Add a new option to control the LTO mode. The default setting of LTO
varies across the different kernel branches, and this flag should only
be specified to *change* an existing setting. If LTO wasn't enabled
to begin with, CFI won't be automatically enabled with this option.

LTO=full enables full LTO, which is traditional LTO. This will make
kernel builds take a long time, but the resulting binaries will be more
optimized. It is not recommended for incremental builds.

LTO=thin enables ThinLTO, if supported. This will make kernel builds
take a little longer, but the resulting binaries will still be optimized
when compared to no LTO, and importantly features like CFI will remain
enabled.

LTO=none disables LTO if it was enabled. This will make kernel builds
run more quickly, but the resulting binaries will not be link-time
optimized. Features like CFI will be implicitly disabled.

This change modifies SKIP_MRPROPER=1 to automatically enable LTO=thin,
as it is expected that this is the primary way most developers will
interact with incremental builds using build.sh. The explicit "LTO"
option is provided for integration with other tooling, or for
experiments.

Bug: 167259389
Signed-off-by: Alistair Delva <adelva@google.com>
Change-Id: Ic45e8e244265fd7e5876f568b143c07a50f4184e