build.sh: Check for hermetic build to use the correct cpio argument
We need to check for hermetic builds in order to determine which cpio
arguments to use. In toybox cpio, --no-preserve-owner is a valid command
line switch for the create i.e. copy-out mode. It causes toybox to set
uid/gid to 0 for all directory entries. This is equivalent the command
line argument -R +0:+0 in GNU cpio. Keep in mind that, in GNU cpio,
--no-preserve-owner means something else and is only valid in copy-in
and copy-pass modes.

Bug: 183088951
Fixes: ba0bfc2a0177 ("build.sh: set all files in initramfs.img to root:root")
Change-Id: I9c1c8179bd08dc9a909f7ed4a9d605d3a772d9d8
Signed-off-by: Will McVicker <willmcvicker@google.com>