build.sh: Always run syncconfig
When running with a few more integration tests, I discovered that
include/config/auto.conf was not in sync with a restore .config.before.

If config.before was resotred, then Kbuild would have auto.con based on
the original defconfig and only knows to update auto.conf based on the
timestamp of .config. Since the .config.before was restored with the old
timestamp, auto.conf would not be updated and thus Kbuild would not have
the up-to-date build configuration.

Bug: 179648610
Fixes: c84e95adb48a ("build.sh: Avoid unnecessary touching of the kernel
config by build.sh")
Signed-off-by: Elliot Berman <eberman@quicinc.com>
Change-Id: Iba5aa1b1c5e558d05bbbd49fc333be1e3c18cb5c