Revert "build.sh: Avoid unnecessary touching of the kernel config by build.sh"
This reverts commit c84e95adb48aa1fc774152844776c48e043eb26a.

It is causing issues with symbol trimming, which can be reproduced as
follows:

  $ rm -rf out/
  $ BUILD_CONFIG=common/build.config.gki.aarch64 ./build/build.sh
  [...]
  $ file out/android12-5.10/common/include/generated/autoksyms.h
  out/android12-5.10/common/include/generated/autoksyms.h: ASCII text
  $ BUILD_CONFIG=common/build.config.gki.aarch64 ./build/build.sh
  [...]
  $ file out/android12-5.10/common/include/generated/autoksyms.h
  [...] No such file or directory

This results in the second build being non-trimmed despite
TRIM_NONLISTED_KMI=1, and it makes the STRICT_MODE complain.

The optimization introduced by this commit is highly desirable, but
until we find a good fix, it is preferable to revert temporarily and
keep the build correct.

Bug: 179648610
Signed-off-by: Quentin Perret <qperret@google.com>
Change-Id: Id2927b9980c1db2ff109e1bde97cf8915e0deea0