diff_abi: emit error message to stderr in case of a failure
In case of a non-ABI related error (crash, incompatibility, other),
diff_abi would pipe the error to abi.report and then just reraise. This
hides the actual error message in abi.report, but leaves the build.log
without a clue.

Instead, now reraise with the actual reason emitted as well.

Signed-off-by: Matthias Maennich <maennich@google.com>
Change-Id: I5c913d9c952fb979fdf3815e366783ec1349c98a