abi: extract_symbols: add --ignore-vmlinux option
If set, assume undefined symbols in modules that cannot be satisfied by
other modules' exports, are defined and exported in vmlinux. In
particular, we do not need to look at the vmlinux in that case.

That is helpful if the vmlinux is trimmed and obviously does not (yet)
export the symbols we want to add during extract_symbols.

Bug: 186695435
Suggested-by: Saravana Kannan <saravanak@google.com>
Signed-off-by: Matthias Maennich <maennich@google.com>
Change-Id: I56c0b96564476621753cce23b4efbb84142f99c3