Kleaf: add rule to selectively capture environment variables
On WORKSPACE load, capture some environment variables. They can
translate into some flags and influence the implementation of generic
rules. Eventually we do not want to use them, but as a transition allow
some to be used. The only one captured now is BUILD_CONFIG to pass the
build config file used for what would be the equivalent of calling
build/build.sh with a BUILD_CONFIG.

Example use in any BUILD file if this is loaded to access the value of
BUILD_CONFIG (if set):

  load("@capture_env//:env.bzl", "BUILD_CONFIG")

Bug: 189451175
Signed-off-by: Matthias Maennich <maennich@google.com>
Change-Id: I875d0333de6728bf7129f8fe985342a0ba8bac04