setup_env: consider externally set SOURCE_DATE_EPOCH
SOURCE_DATE_EPOCH is a convention to indicate the source time stamp. It
generally is treated with preferences over actual file time stamps. If
it is set, take it as a source for the KBUILD_BUILD_TIMESTAMP.

For bazel builds, pass through SOURCE_DATE_EPOCH if it is available and
source it from ${ROOT_DIR}/common if that exists.

Bug: 189451175
Signed-off-by: Matthias Maennich <maennich@google.com>
Change-Id: I8c28d73540c0966d369cf79802348c7c447b20a1