Kleaf: kernel_module check kernel_build for depended module
If kernel_module A depends on kernel_module B, the build rule
for A checks that the kernel_build label for A is the same
as the kernel_build label for B.

The setup script that kernel_module returns in KernelEnvInfo
isn't complete; it depends on kernel_build's setup script. However,
we cannot just add kernel_build's setup script to kernel_module's
setup script, because kernel_build's setup script may not be
idempotent.

Hence, double check that modules refer to the same kernel_build tag,
and only run the setup script from kernel_build once.

Test: builds
Bug: 192617687

Change-Id: Ib1378f054f58ba9ccbfb269408bafaf587d4315b