Kleaf: add --debug_scripts command line option
Introduce a build command line flag that allows to run all script
invocations with `set -x` and a trap that executes `date` after every
command. This helps to debug the combined shell scripts as they get
printed how they are executed. The `date` invocations serve as a poor
man's performance analysis.

This is implemented in terms of KernelEnvInfo passing this down to the
dependent rules as prefix of the setup script, which nicely propagates
to everywhere we need it.

This change uses the hard-coded /usr/bin/date as availability of `date`
in `PATH` is not guaranteed at all times and this is a debug option
anyway.

Bug: 194541987
Signed-off-by: Matthias Maennich <maennich@google.com>
Change-Id: Ib2c8512f46342defd436895666a33ff0511f246a