kleaf: handles SOURCE_DATE_EPOCH properly.
When running with SOURCE_DATE_EPOCH=X, bazelrc
pass through by setting --action_env=SOURCE_DATE_EPOCH.
pass the value to a source_date_epoch() target so that
its value is respected.

If SOURCE_DATE_EPOCH environment variable is not set,
the source_date_epoch() target infers it from the .git
directory.

This duplicates the logic of _setup_env.sh, but because
`git` commands can't be executed in sandboxes [1], this logic
has to be duplicated.

[1] https://github.com/bazelbuild/bazel/issues/7742

Test: Build with and without SOURCE_DATE_EPOCH=X
Test: Build with the same SOURCE_DATE_EPOCH=X twice, result
is cached.
Test: Build with different SOURCE_DATE_EPOCH=X twice,
result is not cached; kernel is rebuilt.

Fixes: 194772369

Change-Id: I1969c21930f0794946c800af4d97dd1dcc2d4ee9