kleaf: handles SOURCE_DATE_EPOCH properly.
When running with SOURCE_DATE_EPOCH=X, bazelrc
passes it through by setting --action_env=SOURCE_DATE_EPOCH.
Pass the value to a source_date_epoch() target so that
its value is respected.

If SOURCE_DATE_EPOCH environment variable is not set,
the source_date_epoch() target infers it from the .git
directory.

This duplicates the logic of _setup_env.sh, but because
`git` commands can't be executed in sandboxes [1], this logic
has to be duplicated. (Alternatively, the _env* target may run
without the sandbox, but let's minimize things executed without
a sandbox to keep the build hermetic.)

[1] https://github.com/bazelbuild/bazel/issues/7742

Note: In _setup_env.sh, if the git command fails, SOURCE_DATE_EPOCH
is set to empty string. Similarly, with source_date_epoch() rule,
if the .git directory does not exist or the git command fails,
SOURCE_DATE_EPOCH is also set to empty string.

Test: Build with and without SOURCE_DATE_EPOCH=X
Test: Build with the same SOURCE_DATE_EPOCH=X twice, result
is cached.
Test: Build with different SOURCE_DATE_EPOCH=X twice,
result is not cached; kernel is rebuilt.

Fixes: 194772369

Change-Id: I1969c21930f0794946c800af4d97dd1dcc2d4ee9