kleaf: preserve SOURCE_DATE_EPOCH
When running with --action_env=SOURCE_DATE_EPOCH=X,
pass the value to the *_env target so that it is
preserved in *_env.sh.

This change does NOT solve the problem when
SOURCE_DATE_EPOCH is unset. In _setup_env.sh, it should be set
from the git history of ${ROOT_DIR}/${KERNEL_DIR}, but there's
no .git directory in the Bazel sandbox. Hence, we still need
to work around it by setting to 0.

Test: Build with and without --action_env=SOURCE_DATE_EPOCH=X
Test: Build with the same --action_env=SOURCE_DATE_EPOCH=X twice, result
is cached.
Test: Build with different --action_env=SOURCE_DATE_EPOCH=X twice,
result is not cached; kernel is rebuilt.

Bug: 194772369

Change-Id: I1969c21930f0794946c800af4d97dd1dcc2d4ee9