kleaf: kernel_modules_install: add MODULES_ARCHIVE.
The default output for kernel_modules_install is also changed
from module_staging_dir to modules_archive. This is so that it
can be accepted by dist. The additional files in the module_staging_dir
archive, like Module.symvers, modules.order etc. conflicts with the ones
in the main kernel build, and should be resolved later.

Test: build dist
Bug: 197995714
Change-Id: Ida5df9b4bb24a720148a25cd1f126b147f6aa6e3