Avoid broken llvm-strip for fips140.ko
llvm-strip deviates from GNU binutils strip, and employs a dubious
heuristic to decide whether a SHT_RELA section is dynamic or static,
even if the file type in question (ET_REL) cannot carry dynamic
relocations to begin with.

Since fips140.ko is a ET_REL ELF binary with SHT_RELA sections that
trigger this issue, let's not strip it at all. This is not actually
needed anyway, given that the kernel build scripts avoid putting debug
metadata in fips140.ko for the same reason.

Change-Id: Iba05339827686d07856651fa1b395af4a7c0a115
Signed-off-by: Ard Biesheuvel <ardb@google.com>