kleaf: kernel_build *outs attributes support select
To support select() properly, the argument "outs" in
kernel_build macro cannot be parsed, and it needs to be
passed to _kernel_build directly. Hence, outs (and
for consistency, module_outs and implicit_outs) of _kernel_build
are changed to a string_list.

Then inside _kernel_build_impl, declare_file() of the proper
output file. Because _kernel_build_impl already explicitly
return DefaultInfo(), this is okay.

To continue to support e.g. "kernel_aarch64/vmlinux", filegroups
with output groups are created to point to the correct file:

  kernel_build(name = "kernel_aarch64", out = ["vmlinux"])

becomes

  _kernel_build(name = "kernel_aarch64", ...)
  filegroup(name = "kernel_aarch64/vmlinux",
            srcs = ["kernel_aarch64"],
            output_group = "vmlinux")

See discussion thread:

  https://github.com/bazelbuild/bazel/issues/281

Bug: 202075567
Bug: 202075499
Test: builds

Change-Id: I92d5c7f357d2d69c40039ef35b3e0a25cdb7214f