kleaf: no output_list for _kernel_build to support select
To support select() properly, the argument "outs" in
kernel_build macro cannot be parsed, and it needs to be
passed to _kernel_build directly. Hence, outs (and
for consistency, module_outs and implicit_outs) of _kernel_build
are changed to a string_list.

Then inside _kernel_build_impl, declare_file() of the proper
output file. Because _kernel_build_impl already explicitly
return DefaultInfo(), this is okay.

The downside is that the following becomes invalid:

  kernel_build(name = "kernel", out = ["vmlinux"])
  filegroup(name = "foo", srcs = ["kernel/vmlinux"])

This is because `kernel/vmlinux` is no longer in the explicit output
list.

See discussion thread:

  https://github.com/bazelbuild/bazel/issues/281

Change-Id: I92d5c7f357d2d69c40039ef35b3e0a25cdb7214f