build.sh: add support for KBUILD_EXT_TREE
If set, this flag defines the path to a driver-only kernel tree that can
be compiled with a supported ACK. The difference between using
KBUILD_EXT_TREE and compiling an external module is that an extended
tree is built with the kernel as if it was in-tree versus out-of-tree.

To enable compiling a vendor extended tree, you need to first setup the
vendor extended tree Kconfig. It should source the base kernel's Kconfig
as well as the drivers you'd like to compile with the base kernel. For
example,

  mainmenu "My extended tree"
  source "$(srctree)/Kconfig.base"
  source "$(KBUILD_EXT_TREE)/mydriver/Kconfig"

After configuring the vendor extended tree, you just need to set
KBUILD_EXT_TREE=/path/to/extended/tree in your BUILD_CONFIG or on the
command line with running build.sh.

Bug: 206700303
Signed-off-by: Will McVicker <willmcvicker@google.com>
Change-Id: I91287debe6aac6be08cd314afd6191cd9818148a