build.sh: add support to package a dtbo.img
Since packaging the dtbo.img requires the tool mkdtimg or mkdtboimg.py
which are under the Apache v2 licence, we can't upstream the packaging
part. So let's add support for this in build.sh since (almost?) all
Android devices use a dtbo.img.

To enable the packaging, you need to set the following flags in your
BUILD_CONFIG:

  MKDTIMG_DTBOS=<list of *.dtbo(s) with respect to the OUT_DIR>
  MKDTIMG_FLAGS=<optional list of flags>

The resulting dtbo.img will be included in the DIST_DIR.

Bug: 204125487
Signed-off-by: Will McVicker <willmcvicker@google.com>
Change-Id: I6ea945a99229af885213d4c0f0439b2ece3b1603