kleaf: support downloading artifacts
Define download_filegroup, that is similar to a filegroup, but
downloads artifacts from ci.android.com.

Because there's nothing "kernel" about it, the rule
is not prefixed with "kernel_".

define_prebuilts() defines a few targets named with suffix
`_prebuilt_or_build`, such that:
- If --gki_prebuilt_num is set, download from ci.android.com
- Otherwise build from source.

define_prebuilts() are intended to be used in device kernel
source trees, but it may be used in common/ as well.

Also fix comments of define_common_kernels().

Bug: 206079661
Test: manual
Change-Id: If9bb026afde3777602e780b7ba65900e55c2bade
Signed-off-by: Yifan Hong <elsk@google.com>