Use gettop.sh to determine ROOT_DIR
This is step 1 to migrate build/ to build/kernel.

Use a consistent logic to determine ROOT_DIR:

  $(gettop.sh)

to determine $ROOT_DIR. This works no matter if build/foo.sh
is a real script or a symlink to build/kernel/foo.sh. This works
no matter if build/foo.sh or build/kernel/foo.sh is invoked.

The location of gettop.sh is determined relative to $0.
- For scripts under build/{kernel/,}
  $($(dirname $0)/gettop.sh)
- For scripts under build/{kernel/,}{static_analysis/,kleaf/}
  $($(dirname $(dirname $0))/gettop.sh)

Bug: 204425264
Test: TH

Change-Id: Ia0432c2db932234944d4828a5bf22277eb892d37