kleaf: Do not generate kernel-headers.tar.gz for device kernel_build().
Only generates it for the GKI kernel_build(). Device-specific
kernel_build() does not generate it.

Note: It is ensured to be kept in _additional_artifacts so that
dist rules with kernel_aarch64_additional_artifacts contains
kernel-headers.tar.gz in its DIST_DIR. In other words, a
device kernel build does NOT build kernel-headers.tar.gz, but
uses the kernel-headers.tar.gz from the GKI build.

Test: TH
Bug: 211695803
Change-Id: I9dd1f8bd89f6e54fa4d89ee1dbb05dbfddab404c