RFC: replace include.tar.gz by declared output directory
This replaces the packaged include.tar.gz by a declared output directory
that is subsequently passed to the downstream target.

This is non-functional as the rsync done in downstream rules makes the
files read-only (which is actually good to catch if downstream targets
mess with those files), but the build stumbles upon that as the
downstream kernel_build rule indeed tries to write to some of those
files (even though it should not).

Uploading as an RFC for discussion.

Bug: 211694935
Signed-off-by: Matthias Maennich <maennich@google.com>
Change-Id: I31d23c3b2955b4de33ca7764087f45b7824001bd