kleaf: support TRIM_NONLISTED_KMI
Create the kleaf counterpart of TRIM_NONLISTED_KMI.
If it is true, trim the output according to KMI symbol list.

In ACK:
- in android-mainline, there are no KMI symbol list.
- in android13-5.x, there are KMI symbol list, and TRIM_NONLISTED_KMI is
  true.

Hence, in common_kernels.bzl, we directly set trim_nonlisted_kmi to be
the value of the existence of the KMI symbol list. If this changes in
the future (e.g. we have the KMI symbol list but we don't want to trim),
then define_common_kernels needs another argument.

Test: build dist
Test: delete a symbol then
      prebuilts/clang/host/linux-x86/clang-r416183b/bin/llvm-nm --defined-only \
           vmlinux | grep ksymtab | cut -d " " -f 3
      shows that the symbol is no longer exported.
Bug: 215745244

Change-Id: I33dbebfd6400c93f600b94b51d2478d3e92387f8