kleaf: ACK build abi_symbollist and trim
In ACK:
- in android-mainline, there are no KMI symbol list.
- in android13-5.x, there are KMI symbol list, and TRIM_NONLISTED_KMI is
  true for aarch64 non debug target.

By default, define_common_kernels() detects the value of
KMI_SYMBOL_LIST +
ADDITIONAL_KMI_SYMBOL_LISTS **NOT** by loading the build
configs, but by glob()-ing for "android/abi_gki_aarch64*".
If these files exist, it assumes that KMI_SYMBOL_LIST and
ADDITIONAL_KMI_SYMBOL_LISTS are defined.

We directly set trim_nonlisted_kmi to be
the value of the existence of the KMI symbol lists for aarch64
non debug target. If this changes in
the future (e.g. we have the KMI symbol list but we don't want to trim),
then define_common_kernels needs another argument.

The aforementioned behavior can be overriden by the
`kmi_config` argument in define_common_kernels.

Also define a kernel_aarch64_notrim target that is the
same as kernel_aarch64, but does not trim. If
kernel_aarch64 does not trim either (e.g. on mainline branch),
kernel_aarch64_notrim is an alias to kernel_aarch64.

Test: bazel run //common:kernel_dist --dist_dir=out/mydist
Test: delete a symbol then
      prebuilts/clang/host/linux-x86/clang-r416183b/bin/llvm-nm --defined-only \
           vmlinux | grep ksymtab | cut -d " " -f 3
      shows that the symbol is no longer exported.

Bug: 215745244

Change-Id: I654ad1192dfc885da3fee3c5f256aa0082798f9c