kleaf: move *.bzl to *_impl.bzl
As part of the build/kleaf to build/kernel/kleaf migration,
there's a period of time when some projects loads from
build/kleaf/*.bzl, while others load from
build/kernel/kleaf/*.bzl. To ensure consistency, the first
step is to move all symbols to an implementation extension,
then write a proxy to load from the appropriate place.

This does not apply to constants.bzl because they are just
constants. It is okay to load from either build/kleaf or
build/kernel/kleaf; they will be the same.

Test: none
Bug: 217630659

Change-Id: I228e941415dc84d582d9c91e197d831e8dd87aee