kleaf: check toolchain_version for kernel_filegroup
If a file named toolchain_version exists for a
kernel_filegroup, when the kernel_filegroup is
used as the base_kernel of another kernel_build rule,
the toolchain_version file is checked against the toolchain_version
of the kernel_build rule.

If there are no files named toolchain_version, a warning
message is printed to say that no check is performed.

Test: manual
Bug: 213939521

Change-Id: Ie7ebadb24e2d323c3cab87e483d2db0b84f8b913