kleaf: Merge _KernelBuildAspectInfo into _KernelModuleDevInfo
This deletes the _kernel_build_aspect.

The aspect was created because the modules_prepare
target was technically not an output of the
_kernel_build target, but just an attribute that
the _kernel_build carries so we can infer the other targets
defined in the kernel_build macro.

However, with the aspect, the _kernel_build rule had
too many interfaces it implements:
- _KernelBuildInfo
- _KernelEnvInfo
- _kernel_build_aspect
- _KernelModuleDevInfo

This makes it hard to replace kernel_build targets with kernel_filegroup
targets, because kernel_filegroup needs to implement all
these interfaces as well.

Delete the _kernel_build_aspect interface and merge it into
_KernelBuildInfo. This allows us to replace kernel_build
with kernel_filegroup in the future.

Test: TH
Bug: 218761417

Change-Id: Ia4f5b3e6e25cba5ef65fc413c830d22f62f50ecf